requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
numpy==1.26.0
